+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,7 @@
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
+            android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,12 +28,15 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
             android:name=".LifeCounterActivity"
+            android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".CardListActivity" >
-        </activity>
         <activity
             android:name=".SettingsActivity"
             android:label="@string/title_activity_settings" >
+        </activity>
+        <activity android:name=".ResultActivity"
+            android:screenOrientation="portrait"
+            >
         </activity>
     </application>
