+++ app/src/main/res/layout/activity_result.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context="com.card_database.mtg.carddatabasemtg.ResultActivity">
+
+    <Button
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="BACK"
+        android:id="@+id/button"
+        android:enabled="true"
+        android:background="#ffffff" />
+
+    <ListView
+        android:layout_width="wrap_content"
+        android:layout_height="465dp"
+        android:id="@+id/listRes"
+        android:layout_gravity="left|top"
+        android:layout_below="@+id/button" />
+
+</RelativeLayout>
