+++ app/src/main/res/layout/fragment_item.xml
@@ -51,7 +51,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceMedium"
-        android:text="Medium Text"
         android:id="@+id/fragment_power_toughness"
         android:layout_below="@+id/fragment_text"
         android:layout_alignParentEnd="true"
